requirements.txt
<<<<<<< Updated upstream
PySide6>=6.5.0
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
ffmpeg-python>=0.2.0
pillow>=10.0.0
gtts>=2.4.0
elevenlabs>=0.2.0
psutil>=5.9.0

# Additional AI providers
anthropic>=0.7.0
azure-cognitiveservices-speech>=1.30.0
stability-sdk>=0.8.0
gtts>=2.4.0
pydub>=0.25.1 
=======
PySide6>=6.5.0
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
ffmpeg-python>=0.2.0
pillow>=10.0.0
gtts>=2.4.0
elevenlabs>=0.2.0
psutil>=5.9.0
openai-whisper>=20231117

# Additional AI providers
anthropic>=0.7.0
azure-cognitiveservices-speech>=1.30.0
stability-sdk>=0.8.0
gtts>=2.4.0
pydub>=0.25.1

# Task Queue & Message Broker for Async Processing
celery==5.4.0
redis==5.0.5
>>>>>>> Stashed changes
